requirements.txt
pillow
pyinsane2
